sc - properly rename folder with capitalization, move App files back to src
