+++ frontend/src/App.js
@@ -1,15 +1,15 @@
 import React from "react";
 import "./App.css";
-import Loading from "../components/Loading/Loading";
-// import AppNavbar from "../components/Nav/AppNavbar";
+import Loading from "components/Loading/Loading";
+import AppNavbar from "components/Nav/AppNavbar";
 import { useAuth0 } from "@auth0/auth0-react";
 import { Container } from "react-bootstrap";
 import { Route, Switch } from "react-router-dom";
-import AppFooter from "../components/Footer/AppFooter";
-import Home from "../pages/Home/Home";
-import Profile from "../pages/Profile/Profile";
-import PrivateRoute from "../components/Auth/PrivateRoute";
-import TodoList from "../pages/Todos/Todos";
+import AppFooter from "components/Footer/AppFooter";
+import Home from "pages/Home/Home";
+import Profile from "pages/Profile/Profile";
+import PrivateRoute from "components/Auth/PrivateRoute";
+import TodoList from "pages/Todos/Todos";
 
 function App() {
   const { isLoading } = useAuth0();
@@ -20,7 +20,7 @@ function App() {
 
   return (
     <div className="App">
-      {/* <AppNavbar /> */}
+      <AppNavbar />
       <Container className="flex-grow-1 mt-5">
         <Switch>
           <Route path="/" exact component={Home} />
